Detects whether an image uses the rule of thirds
uses openCV to detect objects in an image, then checks to see if those objects intersect the four points of interest in the rule of thirds (the rule of thirds is a composition practice in photography). It also checks if the objects run parallel. Returns a list of binary numbers for each object, 1 meaning it intersected or is running parallel, 0 meaning not. I still have to find a way to calculate the probability of the image actually using the rule of thirds instead of just returning a list of ones and zeroes. I also need to improve image detection if possible. Some images don't work well or at all.